Appends a slash to view URLs to prevent the Multimedia plugin from converting to embed code.
MiddMedia "view/download" URLs use the filename as a parameter, so while the
URL may end in ".mp4", it is actually an HTML page. The Multimedia plugin sees
the ".mp4" and makes embedded object markup, but since there is no video data
at the URL the embedded object is broken.